# refactor: Preallocate result in TryParseHex to avoid resizing