@paplorinc 2024-02-27
Thanks for the review @Empact>
We could do without, that's why it's in a separate commit. It's very low risk however and makes the code more readable. If you think maintainability is risky, I can add more tests - what do you think?