@paplorinc 2024-02-27
I have received contradictory comments, so I've removed the last commit to simplify this change by avoiding the resize, hoping this satisfies everyone.