@Empact 2024-02-27
> We could do without, that's why it's in a separate commit. It's very low risk however and makes the code more readable. If you think maintainability is risky, I can add more tests - what do you think?

@paplorinc note that Bitcoin Core is under different constraints than a typical software project. We are review-constrained and change-skeptical in part because we have to be on guard against adversarial changes and also because we are appreciative of the importance of the codebase to the Bitcoin ecosystem.

That said, refactoring and cleanup are important to the long-term health of the project. One way to be respectful of review time is to make focused, well-motivated pull requests, that do not mix various motivations in the same PR.

See also: https://github.com/bitcoin/bitcoin/blob/master/CONTRIBUTING.md#refactoring